svm: create program loader module
