+++ svm/src/lib.rs
@@ -5,6 +5,7 @@ pub mod account_loader;
 pub mod account_overrides;
 pub mod account_rent_state;
 pub mod message_processor;
+pub mod program_loader;
 pub mod transaction_account_state_info;
 pub mod transaction_error_metrics;
 pub mod transaction_processing_callback;
